Refactor `AssetLoader`
`AssetLoader` is now a singleton.  Asset are queued for load
when the game engine is started and a loading scene is displayed.

Once all assets have been loaded, the map is parsed and
constructed and is then loaded via a World scene.

There's still some work to do on the sprites themselves
but that'll come in a different push.